requirements.txt
fastapi
uvicorn
python-multipart
nltk
transformers
torch
sacremoses
reportlab
pydantic
python-docx

